M3-3195 [SDK] Change: NodeBalancer types (#6411)
* Move extended NB node type to features/nb/types, clean up interfaces

* More NB type adjustment

* Copy nb files to new location
