+++ packages/manager/src/features/NodeBalancers/NodeBalancerConfigPanel.tsx
@@ -1,4 +1,3 @@
-import { NodeBalancerConfigNodeFields } from '@linode/api-v4/lib/nodebalancers';
 import { APIError } from '@linode/api-v4/lib/types';
 import * as React from 'react';
 import ActionsPanel from 'src/components/ActionsPanel';
@@ -23,6 +22,7 @@ import TextField from 'src/components/TextField';
 import Toggle from 'src/components/Toggle';
 import { getErrorMap } from 'src/utilities/errorUtils';
 import NodeBalancerConfigNode from './NodeBalancerConfigNode';
+import { NodeBalancerConfigNodeFields } from './types';
 
 type ClassNames =
   | 'divider'
