+++ packages/manager/src/features/NodeBalancers/NodeBalancerDetail/NodeBalancerConfigurations.tsx
@@ -6,7 +6,6 @@ import {
   getNodeBalancerConfigs,
   NodeBalancerConfig,
   NodeBalancerConfigNode,
-  NodeBalancerConfigNodeFields,
   updateNodeBalancerConfigNode
 } from '@linode/api-v4/lib/nodebalancers';
 import { APIError, ResourcePage } from '@linode/api-v4/lib/types';
@@ -49,6 +48,7 @@ import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
 import NodeBalancerConfigPanel from '../NodeBalancerConfigPanel';
 import { lensFrom } from '../NodeBalancerCreate';
+import { NodeBalancerConfigNodeFields } from '../types';
 import {
   clampNumericString,
   createNewNodeBalancerConfig,
